refactor(library): Extract component to represent Police and Crime Commissioner elections
