+++ uk_election_timetables/elections/__init__.py
@@ -4,3 +4,4 @@ from uk_election_timetables.elections.greater_london_assembly import *
 from uk_election_timetables.elections.northern_ireland_assembly import *
 from uk_election_timetables.elections.local import *
 from uk_election_timetables.elections.uk_parliament import *
+from uk_election_timetables.elections.police_and_crime_commissioner import *
